refactor(etl): extract prefect import guard into radiologist.etl.optional
radiologist.cli.optional.require() reached into
radiologist.etl.prefect_pipelines._PREFECT_AVAILABLE, a private
business-logic-module internal with no contract, unlike the dedicated
optional.py sentinel modules registry and inference already have
(_wandb, _uvicorn) that require() reaches into instead.

Move the prefect try/except import guard (flow/task/artifact stand-ins
and the _PREFECT_AVAILABLE/_PREFECT_IMPORT_ERROR flags) into a new
radiologist.etl.optional module mirroring that pattern.
prefect_pipelines.py now imports from it instead of defining it, and
require() reaches into the same kind of seam for every group.

Found during PR #178's post-implementation code review.
