+++ radiologist-cli/src/radiologist/cli/optional.py
@@ -66,7 +66,7 @@ def require(extra: str) -> ModuleType:
         RuntimeError: When the module is absent, or when its feature-level
             sentinel is unavailable (``registry`` ->
             ``radiologist.registry.optional._wandb``, ``etl`` ->
-            ``radiologist.etl.prefect_pipelines._PREFECT_AVAILABLE``,
+            ``radiologist.etl.optional._PREFECT_AVAILABLE``,
             ``inference`` -> module importability only), naming
             ``pip install 'radiologist-cli[<extra>]'``.
     """
@@ -76,9 +76,9 @@ def require(extra: str) -> ModuleType:
         raise RuntimeError(hint)
 
     if extra == "etl":
-        from radiologist.etl import prefect_pipelines
+        from radiologist.etl import optional as etl_optional
 
-        if not prefect_pipelines._PREFECT_AVAILABLE:
+        if not etl_optional._PREFECT_AVAILABLE:
             raise RuntimeError(hint)
     elif extra == "registry":
         from radiologist.registry import optional as registry_optional
